feat(java): use AbstractObjectSerializer to handle common bean / pojo / record copy
