+++ java/fury-core/src/main/java/org/apache/fury/Fury.java
@@ -1239,7 +1239,6 @@ public final class Fury implements BaseFury {
     return deserializeJavaObjectAndClass(buf);
   }
 
-  @SuppressWarnings("unchecked")
   @Override
   public <T> T copy(T obj) {
     if (obj == null) {
