+++ java/fury-core/src/main/java/org/apache/fury/serializer/CompatibleSerializerBase.java
@@ -19,6 +19,7 @@
 
 package org.apache.fury.serializer;
 
+import java.lang.invoke.MethodHandle;
 import org.apache.fury.Fury;
 import org.apache.fury.memory.MemoryBuffer;
 
@@ -26,9 +27,13 @@ import org.apache.fury.memory.MemoryBuffer;
  * Base class for compatible serializer. Both JIT mode serializer and interpreter-mode serializer
  * will extend this class.
  */
-public abstract class CompatibleSerializerBase<T> extends Serializer<T> {
+public abstract class CompatibleSerializerBase<T> extends AbstractObjectSerializer<T> {
   public CompatibleSerializerBase(Fury fury, Class<T> type) {
     super(fury, type);
+  }
+
+  public CompatibleSerializerBase(Fury fury, Class<T> type, MethodHandle constructor) {
+    super(fury, type, constructor);
   }
 
   public T readAndSetFields(MemoryBuffer buffer, T obj) {
